revert back to -u for dolphin user directory flag
this is so dumb.  I thought it was a Mac vs Linux thing, but instead, it
was a v4 vs v5 thing.  how stupid do you have to be to change the
command line flags between versions?  like, thanks for doing it at a
major version number, but seriously?  why the fuck do you care?  I hate
you, Dolphin-at-the-terminal.  you are the worst