trying really hard to make the frame dump flexible
okay, I thought earlier I knew why I was getting two files on my Mac but
only one on Ubuntu.  I thought it had to do with some setting causing
the movie to resize on startup.  I have no clue if this is the case or
not.

Solution!  Just find the largest file in the folder and use that.