pass melee iso into record avi script
instead of hard-coding it in the script file

though get this!  for some reason I cannot fathom, make just refuses to
see that prerequisite with spaces in it.  I tried quoting the argument,
tried adding all kinds of backslashes, tried random stuff like putting
it all on one line … nothing.

or, it’s like … make can see it as a target up above, no problems.  but
it doesn’t think the pattern rule ever matches.  because of the melee
iso line (I tried all kinds of deleting and inserting lines … it’s
definitely the melee iso).  the only thing I found that fixed it was
replacing the spaces with some other character.  so, underscores it is