use secondary expansion for the pattern stems
I don't understand this at all.  By trial and error, I've found the $$*
works as opposed to a plain %.  On my laptop the % worked fine, but
whatever, on server ubuntu with a higher version of make it doesn't work
anymore.  I'm sure somebody can explain that to me, at some point, but I
just don't care.