static pattern rule -> implicit rule
there’s not much point to this change.  in fact, I may be introducing a
bug just because I don’t understand what it was doing in the first
place.  but hey, it’s my project, I can make pointless aesthetic changes
because I feel like it.

like, literally, the only reason this is here is because I feel the
static pattern targets add no value whatsoever.  it’ll change the error
message I get, I think, but I don’t see why I would want to get an error
message that this step failed, as opposed to the standard make message
that it just couldn’t find a matching chain of implicit rules