implicit rule -> static pattern rule
this, on the other hand, is a very important change.  per the comment in
the makefile … with this change I get better error messages.  if I’m
missing the iso, Make will notice that, try to build it, and spit out
the iso error message.  without this static pattern spec, Make would
simply give up because it couldn’t find a satisfactory chain of implicit
rules and complain “no rule matches” with is utterly stupid but whatever