enable dirs as an undocumented target
ugh, my tricks with using directories as order-only prereqs doesn’t
really cover all my use cases.  like, if I go to build a specific video
or something, it bypasses the high-level rules that have the order-only
in them, and then make complains that it can’t find a file or something.

I don’t get it, but meh.