BUG FIX don’t double-count falcon images in falco’s keras folder
OMG so this is why my model can’t learn the difference between falco and
falcon.  the grep was picking up anything with “falco” in it … so things
like “falcon” also got picked up and thrown in the falco bucket.

the fix takes advantage of the naming convention that puts an underscore
after the character name.